*.mk: x86_64 MACHINE is treated as if it were amd64 (?)
I was't able to find this information explicitly, but it seems MACHINE in the FreeBSD build system refers to the build system.
On Linux 'uname -p' and 'uname -m' both return 'x86_64', so in the context of MACHINE* we do for x86_64 what is done for amd64.

Note: THIS MIGHT BE WRONG!

Signed-off-by: Eddy Petrișor <eddy.petrisor@gmail.com>